Update fix AdminHomeController
