--- Source/AirlinesReservationSystem/AirlinesReservationSystem/Areas/Admin/Controllers/HomeController.cs
@@ -1,17 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Web;
-using System.Web.Mvc;
-
-namespace AirlinesReservationSystem.Areas.Admin.Controllers
-{
-    public class HomeController : Controller
-    {
-        // GET: Admin/Home
-        public ActionResult Index()
-        {
-            return View();
-        }
-    }
-}
