+++ Source/AirlinesReservationSystem/AirlinesReservationSystem/Areas/Admin/Views/AdminAirlines/Index.cshtml
@@ -1,0 +1,46 @@
+﻿@model IEnumerable<AirlinesReservationSystem.Airline>
+
+@{
+    ViewBag.Title = "Index";
+    Layout = "~/Views/Shared/_AdminLayout.cshtml";
+}
+
+<h2>Index</h2>
+
+<p>
+    @Html.ActionLink("Create New", "Create")
+</p>
+<table class="table">
+    <tr>
+        <th>
+            @Html.DisplayNameFor(model => model.AirlineName)
+        </th>
+        <th>
+            @Html.DisplayNameFor(model => model.LogoImage)
+        </th>
+        <th>
+            @Html.DisplayNameFor(model => model.Slogan)
+        </th>
+        <th></th>
+    </tr>
+
+@foreach (var item in Model) {
+    <tr>
+        <td>
+            @Html.DisplayFor(modelItem => item.AirlineName)
+        </td>
+        <td>
+            @Html.DisplayFor(modelItem => item.LogoImage)
+        </td>
+        <td>
+            @Html.DisplayFor(modelItem => item.Slogan)
+        </td>
+        <td>
+            @Html.ActionLink("Edit", "Edit", new { id=item.AirlineId }) |
+            @Html.ActionLink("Details", "Details", new { id=item.AirlineId }) |
+            @Html.ActionLink("Delete", "Delete", new { id=item.AirlineId })
+        </td>
+    </tr>
+}
+
+</table>
